docs(examples): make doc examples load page specific #79
