+++ demo/index.html
@@ -32,12 +32,18 @@
     <script type="module">
       import 'https://unpkg.com/marked@latest/marked.min.js';
       import 'https://unpkg.com/prismjs@latest/prism.js';
-      fetch('/demo/demo.md')
+      import { Examples } from './index.js';
+      fetch('/demo/index.md')
         .then((response) => response.text())
         .then((text) => {
           const rawHtml = marked.parse(text);
           document.querySelector('main').innerHTML = rawHtml;
           Prism.highlightAll();
+        })
+        .then(() => {
+          const examples = new Examples();
+
+          examples.initialize();
         });
     </script>
 
@@ -48,6 +54,5 @@
     <script src="https://unpkg.com/@aurodesignsystem/auro-input@latest/dist/auro-input__bundled.js" type="module"></script>
     <script src="https://unpkg.com/@alaskaairux/auro-button@latest/dist/auro-button__bundled.js" type="module"></script>
     <script type="module" src="../src/auro-combobox.js"></script>
-    <script type="module" src="./examples.js"></script>
   </body>
 </html>
