+++ package.json
@@ -20,21 +20,22 @@
   },
   "dependencies": {
     "@alaskaairux/icons": "^4.13.0",
+    "@aurodesignsystem/auro-menu": "^3.9.1",
     "chalk": "^4.1.2",
     "lit-element": "^2.5.1",
     "lit-html": "^1.4.1"
   },
   "peerDependencies": {
     "@alaskaairux/design-tokens": "^3.6.0",
-    "@aurodesignsystem/auro-dropdown": "^2.9.0",
-    "@aurodesignsystem/auro-input": "^2.9.5",
+    "@aurodesignsystem/auro-dropdown": "^2.9.1",
+    "@aurodesignsystem/auro-input": "^2.10.0",
     "@aurodesignsystem/webcorestylesheets": "^4.0.0"
   },
   "devDependencies": {
-    "@alaskaairux/design-tokens": "^3.9.0",
+    "@alaskaairux/design-tokens": "^3.11.0",
     "@aurodesignsystem/auro-menu": "^3.9.1",
     "@aurodesignsystem/eslint-config": "^1.3.0",
-    "@aurodesignsystem/webcorestylesheets": "^4.0.0",
+    "@aurodesignsystem/webcorestylesheets": "^4.1.1",
     "@commitlint/cli": "^16.3.0",
     "@commitlint/config-conventional": "^16.2.4",
     "@open-wc/testing": "^3.1.5",
